refactor: nestjs auth refactor
