+++ packages/utils-nest/src/auth/guards/cred-guard.ts
@@ -1,19 +1,8 @@
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common'
-import { ApiKeyProvider } from './api-key-provider'
-import { TokenProvider } from './token-provider'
+import { ApiKeyProvider } from './../api-key-provider'
+import { TokenProvider } from './../token-provider'
 import { Request } from 'express'
-import { LoginIdentity } from '@reapit/connect-session'
-import { ApiKeyModel } from '@reapit/api-key-verify'
-
-type LoginCredsType = {
-  type: 'jwt'
-} & LoginIdentity
-
-type ApiKeyCredsType = {
-  type: 'api-key'
-} & Omit<ApiKeyModel, 'expired'>
-
-export type CredsType = LoginCredsType | ApiKeyCredsType
+import { CredsType } from './cred-types'
 
 @Injectable()
 export class CredGuard implements CanActivate {
