+++ packages/utils-nest/src/auth/index.ts
@@ -1,7 +1,5 @@
 export * from './module'
-export * from './ownership'
-export * from './cred-guard'
-export * from './admin-guard'
-export * from './cred-decorator'
+export * from './ownership-provider'
+export * from './guards'
 export * from './api-key-invoke-config'
-export * from './is-admin'
+export * from './helpers'
